requirements.txt
absl-py==1.4.0
array-record==0.4.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
blurhash==1.1.4
boost==0.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
colorama==0.4.6
configobj==5.0.8
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dlib==19.22.99
dm-tree==0.1.8
etils==1.5.2
exceptiongroup==1.2.0
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.45.1
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipython==8.18.0
jedi==0.19.1
Jinja2==3.1.2
JsonForm==0.0.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
JsonSir==0.0.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
Mastodon.py==1.8.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
pip==23.3.1
promise==2.3
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyCameraList==1.0.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
Python-EasyConfig==0.1.7
python-magic-bin==0.4.14
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
Resource==0.2.1
rpds-py==0.13.1
rsa==4.9
scipy==1.11.4
setuptools==68.0.0
six==1.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
toml==0.10.2
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.13.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.14.1
xlwings==0.30.12
zipp==3.17.0
